taskflow/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
asyncpg==0.29.0
bcrypt==4.1.3
python-jose[cryptography]==3.3.0
pydantic[email]==2.7.4
pydantic-settings==2.3.4
pytest==8.2.2
httpx==0.27.0
pytest-asyncio==0.23.7
